fix(core): capability override glyphs, layout truncation, table head parsing
capability_override.rs: Add three new CapabilityOverride fields for glyph
support control — unicode_box_drawing, unicode_emoji, double_width — with
builder methods and preset updates (dumb/full/minimal_modern). Allows
overriding glyph detection without changing the full terminal profile.

input_parser.rs: Add explicit block scope and debug_assert!(f_num <= 24)
guard to the Kitty keyboard F1-F24 range handler, preventing silent
wrapping on malformed key codes.

lib.rs (ftui-layout): Fix integer truncation bug in Flex spacing
distribution — remainder is now kept as usize and compared with the loop
index as usize, preventing incorrect extra-pixel distribution when item
count exceeds u16::MAX.

diff.rs (ftui-render): Add is_coverage_run() helper and relax the p95
performance budget from 500us to 3000us when running under coverage
instrumentation (FTUI_COVERAGE, LLVM_PROFILE_FILE, CARGO_LLVM_COV).

markdown.rs (ftui-extras): Fix table header flag leak — TableHead end tag
now explicitly clears in_head after committing the header row, preventing
body rows from being misclassified as headers. Simplify TableRow end tag
to use in_head directly.

+++ crates/ftui-layout/src/lib.rs
@@ -581,8 +581,9 @@ impl Flex {
                         let count = sizes.len() - 1; // usize
                         // Use usize division to prevent overflow/panic
                         let base = (leftover as usize / count) as u16;
-                        let rem = (leftover as usize % count) as u16;
-                        let extra = base + if (i as u16) < rem { 1 } else { 0 };
+                        let rem = leftover as usize % count;
+                        // Compare i (usize) with rem (usize) to avoid truncation for large i
+                        let extra = base + if i < rem { 1 } else { 0 };
                         current_pos = current_pos.saturating_add(extra);
                     }
                 }
